split up parser tests into multiple files
